fixing linting issues
